+++ lib/structure/content.php
@@ -20,12 +20,6 @@ add_filter(
 	'__return_true'
 );
 
-/* Display author box on archive pages */
-add_filter(
-	'get_the_author_genesis_author_box_archive',
-	'__return_true'
-);
-
 /**
  * Modify the size of the Gravatar in the author box
  *